requirements.txt
faiss-cpu>=1.7.0
torch>=1.9.0
transformers>=4.20.0
numpy>=1.21.0
Pillow>=8.0.0
tqdm>=4.60.0
google-generativeai>=0.3.0